ConfluxAI_hybrid/requirements.txt
# FastAPI and web framework dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data validation and serialization
pydantic>=2.5.0

# Vector embeddings and similarity search
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.26.0

# PDF processing
PyPDF2>=3.0.1
pypdf>=3.17.4

# Image processing and OCR
Pillow>=10.1.0
pytesseract>=0.3.10

# Document processing
python-docx>=1.1.0
openpyxl>=3.1.2
pandas>=2.1.4

# Text processing
nltk>=3.8.1

# Logging and configuration
python-dotenv>=1.0.0

# HTTP client for testing
httpx>=0.25.2

# Development dependencies (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Phase 2 Enhancement Dependencies
# Advanced PDF processing
pdfplumber>=0.9.0
tabula-py>=2.7.0
camelot-py[cv]>=0.10.1

# Advanced image processing and computer vision
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Hybrid search capabilities
rank-bm25>=0.2.2

# Caching and background processing
redis>=4.5.0
celery>=5.3.0

# Additional file format support
beautifulsoup4>=4.12.0
python-pptx>=0.6.21
markdown>=3.4.0

# Performance and monitoring
psutil>=5.9.0
prometheus-client>=0.16.0

# Phase 3 AI Dependencies
# AI/ML Core Libraries
transformers>=4.35.0
torch>=2.1.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.25.0

# Language Detection
langdetect>=1.0.9

# Advanced NLP
spacy>=3.7.0
scikit-learn>=1.3.0

# WebSocket Support
fastapi-socketio>=0.0.10
python-socketio>=5.10.0
websockets>=12.0

# Additional Database Support
asyncpg>=0.29.0
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.12.0

# Image and Multi-modal Processing
clip-by-openai>=1.0
